requirements.txt
aniso8601==9.0.1
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
dnspython==1.16.0
Flask==2.0.1
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.9
gunicorn==20.1.0
idna==2.10
importlib-metadata==4.6.0
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
MarkupSafe==2.0.1
numpy==1.21.0
pandas==1.3.0
pymongo==3.11.4
python-dateutil==2.8.1
python-dotenv==0.18.0
pytz==2021.1
requests==2.25.1
scikit-learn==0.24.2
scipy==1.7.0
six==1.16.0
threadpoolctl==2.2.0
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==2.0.1
xgboost==1.4.2
zipp==3.4.1
